wip, close #3252

+++ src/Form.php
@@ -6,6 +6,7 @@ use Closure;
 use Encore\Admin\Exception\Handler;
 use Encore\Admin\Form\Builder;
 use Encore\Admin\Form\Field;
+use Encore\Admin\Form\HasHooks;
 use Encore\Admin\Form\Row;
 use Encore\Admin\Form\Tab;
 use Illuminate\Contracts\Support\Renderable;
@@ -15,7 +16,6 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Input;
 use Illuminate\Support\MessageBag;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Validator;
@@ -72,6 +72,13 @@ use Symfony\Component\HttpFoundation\Response;
  */
 class Form implements Renderable
 {
+    use HasHooks;
+
+    /**
+     * Remove flag in `has many` form.
+     */
+    const REMOVE_FLAG_NAME = '_remove_';
+
     /**
      * Eloquent model of the form.
      *
@@ -88,34 +95,6 @@ class Form implements Renderable
      * @var Builder
      */
     protected $builder;
-
-    /**
-     * Submitted callback.
-     *
-     * @var Closure[]
-     */
-    protected $submitted = [];
-
-    /**
-     * Saving callback.
-     *
-     * @var Closure[]
-     */
-    protected $saving = [];
-
-    /**
-     * Saved callback.
-     *
-     * @var Closure[]
-     */
-    protected $saved = [];
-
-    /**
-     * Callbacks after getting editing model.
-     *
-     * @var Closure[]
-     */
-    protected $editing = [];
 
     /**
      * Data for save to current model from input.
@@ -170,11 +149,6 @@ class Form implements Renderable
      * @var Form\Tab
      */
     protected $tab = null;
-
-    /**
-     * Remove flag in `has many` form.
-     */
-    const REMOVE_FLAG_NAME = '_remove_';
 
     /**
      * Field rows in form.
@@ -325,27 +299,49 @@ class Form implements Renderable
      */
     public function destroy($id)
     {
-        collect(explode(',', $id))->filter()->each(function ($id) {
-            $builder = $this->model()->newQuery();
+        try {
 
-            if ($this->isSoftDeletes) {
-                $builder = $builder->withTrashed();
+            if (($ret = $this->callDeleting()) instanceof Response) {
+                return $ret;
             }
 
-            $model = $builder->with($this->getRelations())->findOrFail($id);
+            collect(explode(',', $id))->filter()->each(function ($id) {
+                $builder = $this->model()->newQuery();
 
-            if ($this->isSoftDeletes && $model->trashed()) {
-                $this->deleteFiles($model, true);
-                $model->forceDelete();
+                if ($this->isSoftDeletes) {
+                    $builder = $builder->withTrashed();
+                }
 
-                return;
+                $model = $builder->with($this->getRelations())->findOrFail($id);
+
+                if ($this->isSoftDeletes && $model->trashed()) {
+                    $this->deleteFiles($model, true);
+                    $model->forceDelete();
+
+                    return;
+                }
+
+                $this->deleteFiles($model);
+                $model->delete();
+            });
+
+            if (($ret = $this->callDeleted()) instanceof Response) {
+                return $ret;
             }
 
-            $this->deleteFiles($model);
-            $model->delete();
-        });
+            $response = [
+                'status'  => true,
+                'message' => trans('admin.delete_succeeded'),
+            ];
 
-        return true;
+        } catch (\Exception $exception) {
+            $response = [
+                'status'  => false,
+                'message' => $exception->getMessage() ?: trans('admin.delete_failed'),
+            ];
+        }
+
+        return response()->json($response);
     }
 
     /**
@@ -379,7 +375,7 @@ class Form implements Renderable
      */
     public function store()
     {
-        $data = Input::all();
+        $data = \request()->all();
 
         // Handle validation errors.
         if ($validationMessages = $this->validationMessages($data)) {
@@ -498,61 +494,6 @@ class Form implements Renderable
     }
 
     /**
-     * Call editing callbacks.
-     *
-     * @return void
-     */
-    protected function callEditing()
-    {
-        foreach ($this->editing as $func) {
-            call_user_func($func, $this);
-        }
-    }
-
-    /**
-     * Call submitted callback.
-     *
-     * @return mixed
-     */
-    protected function callSubmitted()
-    {
-        foreach ($this->submitted as $func) {
-            if ($func instanceof Closure && ($ret = call_user_func($func, $this)) instanceof Response) {
-                return $ret;
-            }
-        }
-    }
-
-    /**
-     * Call saving callback.
-     *
-     * @return mixed
-     */
-    protected function callSaving()
-    {
-        foreach ($this->saving as $func) {
-            if ($func instanceof Closure && ($ret = call_user_func($func, $this)) instanceof Response) {
-                return $ret;
-            }
-        }
-    }
-
-    /**
-     * Callback after saving a Model.
-     *
-     * @return mixed|null
-     */
-    protected function callSaved()
-    {
-        foreach ($this->saved as $func) {
-            if ($func instanceof Closure &&
-                ($ret = call_user_func($func, $this)) instanceof Response) {
-                return $ret;
-            }
-        }
-    }
-
-    /**
      * Handle update.
      *
      * @param int  $id
@@ -562,7 +503,7 @@ class Form implements Renderable
      */
     public function update($id, $data = null)
     {
-        $data = ($data) ?: Input::all();
+        $data = ($data) ?: request()->all();
 
         $isEditable = $this->isEditable($data);
 
@@ -744,7 +685,7 @@ class Form implements Renderable
             unset($input['key']);
         }
 
-        Input::replace($input);
+        request()->replace($input);
 
         return $input;
     }
@@ -770,7 +711,7 @@ class Form implements Renderable
             $input[$column] = $order;
         }
 
-        Input::replace($input);
+        request()->replace($input);
 
         return $input;
     }
@@ -1023,54 +964,6 @@ class Form implements Renderable
         }
 
         return Arr::isAssoc($first);
-    }
-
-    /**
-     * Set after getting editing model callback.
-     *
-     * @param Closure $callback
-     *
-     * @return void
-     */
-    public function editing(Closure $callback)
-    {
-        $this->editing[] = $callback;
-    }
-
-    /**
-     * Set submitted callback.
-     *
-     * @param Closure $callback
-     *
-     * @return void
-     */
-    public function submitted(Closure $callback)
-    {
-        $this->submitted[] = $callback;
-    }
-
-    /**
-     * Set saving callback.
-     *
-     * @param Closure $callback
-     *
-     * @return void
-     */
-    public function saving(Closure $callback)
-    {
-        $this->saving[] = $callback;
-    }
-
-    /**
-     * Set saved callback.
-     *
-     * @param Closure $callback
-     *
-     * @return void
-     */
-    public function saved(Closure $callback)
-    {
-        $this->saved[] = $callback;
     }
 
     /**
@@ -1361,6 +1254,8 @@ class Form implements Renderable
     /**
      * Disable form submit.
      *
+     * @param bool $disable
+     *
      * @return $this
      *
      * @deprecated
@@ -1374,6 +1269,8 @@ class Form implements Renderable
 
     /**
      * Disable form reset.
+     *
+     * @param bool $disable
      *
      * @return $this
      *
@@ -1389,6 +1286,8 @@ class Form implements Renderable
     /**
      * Disable View Checkbox on footer.
      *
+     * @param bool $disable
+     *
      * @return $this
      */
     public function disableViewCheck(bool $disable = true)
@@ -1401,6 +1300,8 @@ class Form implements Renderable
     /**
      * Disable Editing Checkbox on footer.
      *
+     * @param bool $disable
+     *
      * @return $this
      */
     public function disableEditingCheck(bool $disable = true)
@@ -1412,6 +1313,8 @@ class Form implements Renderable
 
     /**
      * Disable Creating Checkbox on footer.
+     *
+     * @param bool $disable
      *
      * @return $this
      */
@@ -1649,7 +1552,8 @@ class Form implements Renderable
      * Setter.
      *
      * @param string $name
-     * @param $value
+     * @param mixed  $value
+     * @return array
      */
     public function __set($name, $value)
     {
